+++ src/launcher/autoattach/luaversion.cpp
@@ -1,0 +1,81 @@
+#include <autoattach/luaversion.h>
+#include <charconv>
+#include <gumpp.hpp>
+
+namespace luadebug::autoattach {
+    lua_version lua_version_from_string(const std::string_view& v) {
+        if (v == "luajit")
+            return lua_version::luajit;
+        if (v == "lua51")
+            return lua_version::lua51;
+        if (v == "lua52")
+            return lua_version::lua52;
+        if (v == "lua53")
+            return lua_version::lua53;
+        if (v == "lua54")
+            return lua_version::lua54;
+        return lua_version::unknown;
+    }
+
+    const char* lua_version_to_string(lua_version v) {
+        switch (v) {
+        case lua_version::lua51:
+            return "lua51";
+        case lua_version::lua52:
+            return "lua52";
+        case lua_version::lua53:
+            return "lua53";
+        case lua_version::lua54:
+            return "lua54";
+        case lua_version::luajit:
+            return "luajit";
+        default:
+            return "unknown";
+        }
+    }
+
+    lua_version get_lua_version(const RuntimeModule& rm) {
+        /*
+            luaJIT_version_2_1_0_beta3
+            luaJIT_version_2_1_0_beta2
+            luaJIT_version_2_1_0_beta1
+            luaJIT_version_2_1_0_alpha
+        */
+        for (void* addr : Gum::SymbolUtil::find_matching_functions("luaJIT_version_2_1_0*", true)){
+            if (rm.in_module(addr))
+                return lua_version::luajit;
+        }
+        auto p = Gum::Process::module_find_symbol_by_name(rm.path, "lua_ident");;
+        const char *lua_ident = (const char *) p;
+        if (!lua_ident)
+            return lua_version::unknown;
+        auto id = std::string_view(lua_ident);
+        using namespace std::string_view_literals;
+        constexpr auto key = "$Lua"sv;
+        if (id.substr(0, key.length()) != key) {
+            return lua_version::unknown;
+        }
+        if (id[key.length()] == ':') {
+            //$Lua: Lua 5.1.*
+            return lua_version::lua51;
+        }
+        id = id.substr("$LuaVersion: Lua 5."sv.length());
+        int patch;
+        auto res = std::from_chars(id.data(), id.data() + 3, patch, 10);
+        if (res.ec != std::errc()) {
+            return lua_version::unknown;
+        }
+        switch (patch) {
+        case 1:
+            return lua_version::lua51;
+        case 2:
+            return lua_version::lua52;
+        case 3:
+            return lua_version::lua53;
+        case 4:
+            return lua_version::lua54;
+        default:
+            return lua_version::unknown;
+        }
+    }
+}
