+++ src/launcher/util/common.hpp
@@ -12,17 +12,3 @@ typedef struct _RuntimeModule {
         return addr >load_address && addr <= (void*)((intptr_t)load_address + size);
     }
 } RuntimeModule;
-
-namespace luadebug::autoattach {
-    enum class lua_version {
-        unknown,
-        luajit,
-        lua51,
-        lua52,
-        lua53,
-        lua54,
-		max,
-    };
-    lua_version lua_version_from_string(const std::string_view& v);
-    const char* lua_version_to_string(lua_version v);
-}
